+++ test/case/env-test/src/entry/index/index.js
@@ -1,0 +1,5 @@
+/* eslint no-console: 0 */
+import './index.scss';
+
+window.mode = process.env.NODE_ENV;
+console.log(`env: ${process.env.NODE_ENV}`);