+++ ui-test/test.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  INIT: true,
+  WATCH: true,
+  ENV: true
+};