+++ yyt.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  default: {
+    src_folders: ['ui-test/test'],
+    custom_commands_path: ['ui-test/commands'],
+    output_folder: false,
+    test_settings: {
+      default: {
+        globals: {
+          asyncHookTimeout : 60000
+        }
+      }
+    }
+  }
+};