+++ history.md
@@ -1,4 +1,8 @@
 # 版本信息
+## 0.5.0 (2019-01-30)
+* [ADD] 支持项目中配置 自己的 `package.json` 和 `webpack.config.js`
+* [ADD] `---anything 123` 会映射到 项目中的 `proccess.env.anything === 123`
+
 ## 0.4.2 (2019-01-03)
 * [FIX] 修复 `multi entry` 时 js 打包 没有独立分开问题